+++ changelog.md
@@ -1,3 +1,16 @@
+## v0.2
+
+  * added `TimeSig` type and associated functions, so time signatures
+    in a `MeasureMap` remember their denominators
+    instead of e.g. converting 6/8 to 3/4
+
+  * `unmakeTempoMap`, `unmakeMeasureMap`
+
+  * relicensed as BSD3
+
+  * note: the only change that necessitated a major version bump
+    was the removal of the `Ord` instance for `TempoMap` and `MeasureMap`
+
 ## v0.1.1.1
 
   * fixed a bug in "extractFirst" (the time value returned was incorrect)